rename some files for consistency
