+++ how-to/src/XSignup.js
@@ -1,11 +1,11 @@
 import React, { useState, useEffect } from 'react'
 import './XApp.css'
-import XForm from './XForm'
+import XSignupForm from './XSignupForm'
 import XSignIn from './XSignIn'
 import XUser from './XUser'
 import axios from "axios"
 import * as yup from "yup"
-import schema from './XformSchema'
+import schema from './XsignupSchema'
 
 const initialFormValues = {
     name: '',
@@ -26,10 +26,10 @@ const initialFormErrors = {
     position: '',
     tos: '\n',
 }
-const initialUsers = [{},{}]
+const initialUsers = [{}]
 const initialDisabled = true
 
-function XApp() {
+function App() {
   
     const [users, setUsers] = useState(initialUsers)
     const [formValues, setFormValues] = useState(initialFormValues)
@@ -112,14 +112,13 @@ function XApp() {
 // debugger
     return (
         <div className="XApp">
-            {/* <XForm 
+            <XForm 
                 values={formValues}
                 change={inputChange}
                 submit={formSubmit}
                 disabled={disabled}
                 errors={formErrors}
-            /> */}
-            <XSignIn />
+            />
             {users.map((user) => {
                 return <XUser key={user.id} details={user} />
             })}
@@ -127,4 +126,4 @@ function XApp() {
     )
 }
 
-export default XApp
+export default App
